--- EssentialDevApp/EssentialDevApp/FeedLoaderPresentationAdapter.swift
@@ -1,36 +1,0 @@
-//
-//  FeedLoaderPresentationAdapter.swift
-//  EssentialDeviOS
-//
-//  Created by Min-Yang Huang on 4/26/25.
-//
-
-import Combine
-import EssentialDev
-import EssentialDeviOS
-
-final class FeedLoaderPresentationAdapter: FeedViewControllerDelegate {
-    private let feedLoader: () -> AnyPublisher<[FeedImage], Error>
-    var presenter: LoadResourcePresenter<[FeedImage], FeedViewAdapter>?
-    var cancellable: Cancellable?
-    
-    init(feedLoader: @escaping () -> AnyPublisher<[FeedImage], Error>) {
-        self.feedLoader = feedLoader
-    }
-    
-    func didRequestFeedRefresh() {
-        presenter?.didStartLoading()
-        
-        cancellable = feedLoader()
-            .dispatchOnMainQueue()
-            .sink(receiveCompletion: { [weak self] completion in
-                switch completion {
-                case .finished: break
-                case let .failure(error):
-                    self?.presenter?.didFinishLoading(with: error)
-                }
-            }, receiveValue: { [weak self] feed in
-                self?.presenter?.didFinishLoading(with: feed)
-            })
-    }
-}
